removed feature that skipped tests enabling them
