+++ dash-spv/Cargo.toml
@@ -81,8 +81,5 @@ name = "dash_spv"
 path = "src/lib.rs"
 
 [features]
-# Gate incomplete mock-dependent tests to avoid "unexpected cfg" warnings.
-skip_mock_implementation_incomplete = []
-
 # Terminal UI feature (off by default, for use by binary only)
 terminal-ui = ["dep:crossterm"]
