+++ dash-spv/tests/block_download_test.rs
@@ -1,11 +1,4 @@
 //! Tests for block downloading on filter match functionality.
-//!
-//! NOTE: This test file is currently disabled due to incomplete mock NetworkManager implementation.
-//! TODO: Re-enable once NetworkManager trait methods are fully implemented.
-
-#![cfg(feature = "skip_mock_implementation_incomplete")]
-
-//! Tests for block downloading on filter match functionality.
 
 use std::collections::HashSet;
 use std::sync::Arc;
@@ -159,7 +152,6 @@ fn create_test_filter_match(block_hash: BlockHash, height: u32) -> FilterMatch {
     }
 }
 
-#[ignore = "mock implementation incomplete"]
 #[tokio::test]
 async fn test_filter_sync_manager_creation() {
     let config = create_test_config();
@@ -171,7 +163,6 @@ async fn test_filter_sync_manager_creation() {
     assert_eq!(filter_sync.pending_download_count(), 0);
 }
 
-#[ignore = "mock implementation incomplete"]
 #[tokio::test]
 async fn test_request_block_download() {
     let config = create_test_config();
@@ -209,7 +200,6 @@ async fn test_request_block_download() {
     assert_eq!(filter_sync.pending_download_count(), 1);
 }
 
-#[ignore = "mock implementation incomplete"]
 #[tokio::test]
 async fn test_duplicate_block_request_prevention() {
     let config = create_test_config();
@@ -233,7 +223,6 @@ async fn test_duplicate_block_request_prevention() {
     assert_eq!(filter_sync.pending_download_count(), 1);
 }
 
-#[ignore = "mock implementation incomplete"]
 #[tokio::test]
 async fn test_handle_downloaded_block() {
     let config = create_test_config();
@@ -264,7 +253,6 @@ async fn test_handle_downloaded_block() {
     assert_eq!(filter_sync.pending_download_count(), 0);
 }
 
-#[ignore = "mock implementation incomplete"]
 #[tokio::test]
 async fn test_handle_unexpected_block() {
     let config = create_test_config();
@@ -281,7 +269,6 @@ async fn test_handle_unexpected_block() {
     assert!(result.is_none());
 }
 
-#[ignore = "mock implementation incomplete"]
 #[tokio::test]
 async fn test_process_multiple_filter_matches() {
     let config = create_test_config();
@@ -332,11 +319,9 @@ async fn test_process_multiple_filter_matches() {
     assert_eq!(filter_sync.pending_download_count(), 3);
 }
 
-#[ignore = "mock implementation incomplete"]
 #[tokio::test]
 async fn test_sync_manager_integration() {}
 
-#[ignore = "mock implementation incomplete"]
 #[tokio::test]
 async fn test_filter_match_and_download_workflow() {
     let config = create_test_config();
@@ -369,7 +354,6 @@ async fn test_filter_match_and_download_workflow() {
     assert!(filter_sync.has_pending_downloads());
 }
 
-#[ignore = "mock implementation incomplete"]
 #[tokio::test]
 async fn test_reset_clears_download_state() {
     let config = create_test_config();
