+++ dash-spv/tests/filter_header_verification_test.rs
@@ -1,12 +1,5 @@
 //! Test to replicate the filter header chain verification failure observed in production.
 //!
-//! NOTE: This test file is currently disabled due to incomplete mock NetworkManager implementation.
-//! TODO: Re-enable once NetworkManager trait methods are fully implemented.
-
-#![cfg(feature = "skip_mock_implementation_incomplete")]
-
-//! Test to replicate the filter header chain verification failure observed in production.
-//!
 //! This test reproduces the exact scenario from the logs where:
 //! 1. A batch of 1999 filter headers from height 616001-617999 is processed successfully
 //! 2. The next batch starting at height 618000 fails verification because the
